+++ dataclass_array/type_parsing_test.py
@@ -25,6 +25,7 @@ from etils.array_types import f32, FloatArray  # pylint: disable=g-multiple-impo
 import pytest
 
 
+_DS = dca.field_utils.DataclassWithShape
 Ray = dca.testing.Ray
 
 
@@ -39,7 +40,10 @@ class Camera(dca.DataclassArray):
     [
         (int, [int]),
         (Ray, [Ray]),
+        (Ray['h w'], [Ray['h w']]),
+        (Ray[..., 3], [Ray[..., 3]]),
         (Union[Ray, int], [Ray, int]),
+        (Union[Ray['h w'], int], [Ray['h w'], int]),
         (Union[Ray, int, None], [Ray, int, None]),
         (Optional[Ray], [Ray, None]),
         (Optional[Union[Ray, int]], [Ray, int, None]),
@@ -55,10 +59,12 @@ def test_get_leaf_types(hint, expected):
     'hint, expected',
     [
         (int, None),
-        (Ray, Ray),
-        (Optional[Ray], Ray),
-        (Union[Ray, Camera], dca.DataclassArray),
-        (Union[Ray, Camera, None], dca.DataclassArray),
+        (Ray, _DS(Ray, '...')),
+        (Ray['h w'], _DS(Ray, 'h w')),
+        (Ray[..., 3], _DS(Ray, '... 3')),
+        (Optional[Ray], _DS(Ray, '...')),
+        (Union[Ray, Camera], _DS(dca.DataclassArray, '...')),
+        (Union[Ray, Camera, None], _DS(dca.DataclassArray, '...')),
         (Union[Ray, int], None),
         (Union[Ray, int, None], None),
         (Union[f32[3, 3], int, None], None),
@@ -72,9 +78,51 @@ def test_get_array_type(hint, expected):
   assert type_parsing.get_array_type(hint) == expected
 
 
+@pytest.mark.parametrize(
+    'hint, expected',
+    [
+        (Ray, _DS(Ray, '...')),
+        (Ray['h w'], _DS(Ray, 'h w')),
+        (Ray[..., 3], _DS(Ray, '... 3')),
+    ],
+)
+def test_from_hint(hint, expected):
+  assert dca.field_utils.DataclassWithShape.from_hint(hint) == expected
+
+
 def test_get_array_type_error():
   with pytest.raises(NotImplementedError):
     type_parsing.get_array_type(Union[Ray, f32[3, 3]])
 
   with pytest.raises(NotImplementedError):
     type_parsing.get_array_type(Union[FloatArray[..., 3], f32[3, 3]])
+
+
+@pytest.mark.parametrize(
+    'hint, expected',
+    [
+        (
+            Ray,
+            dca.array_dataclass._ArrayFieldMetadata(
+                inner_shape_non_static=(),
+                dtype=Ray,
+            ),
+        ),
+        (
+            Ray[..., 3],
+            dca.array_dataclass._ArrayFieldMetadata(
+                inner_shape_non_static=(3,),
+                dtype=Ray,
+            ),
+        ),
+        (
+            Ray['*shape 4 _'],
+            dca.array_dataclass._ArrayFieldMetadata(
+                inner_shape_non_static=(4, None),
+                dtype=Ray,
+            ),
+        ),
+    ],
+)
+def test_type_to_field_metadata(hint, expected):
+  assert dca.array_dataclass._type_to_field_metadata(hint) == expected